sting: change unit of offset to unsigned long

+++ kernel/sting/include/ept_dict.h
@@ -1,6 +1,6 @@
 struct ept_dict_key {
 	ino_t ino;
-	int offset;
+	unsigned long offset;
 };
 
 struct ept_dict_val {
